+++ lib/src/rx/rx_list_impl.dart
@@ -1,40 +1,19 @@
 part of view_model_kit;
 
-abstract class _BaseR {
-  void Function() get _notify;
-}
-
-/// Reactive Value (int, double, string, model(value object), ...)
-class R<V> implements _BaseR {
-  V _value;
-
-  @override
-  late final void Function() _notify;
-
-  R._(this._value);
-
-  V get value => _value;
-
-  set value(V newValue) {
-    _value = newValue;
-    _notify();
-  }
-
-  @override
-  String toString() => "R($_value)";
-}
-
 /// Reactive List
-class RList<V> implements _BaseR {
+class RList<V> extends BaseR<List<V>> {
   final List<V> _list;
 
   @override
-  late final void Function() _notify;
+  late final void Function() _notifyAtStatefulWidget;
 
   RList._(this._list);
 
   /// Read
 
+  /// do not write with this list.
+  @override
+  @protected
   List<V> get value => _list;
 
   V operator [](int index) => _list[index];
@@ -42,7 +21,7 @@ class RList<V> implements _BaseR {
   int get length => _list.length;
 
   /// Write
-
+  ///
   void operator []=(int index, V value) {
     _list[index] = value;
     _notify();
@@ -58,9 +37,16 @@ class RList<V> implements _BaseR {
     _notify();
   }
 
-  void remove(V value) {
-    _list.remove(value);
+  bool remove(Object? value) {
+    final res = _list.remove(value);
     _notify();
+    return res;
+  }
+
+  V removeLast() {
+    final v = _list.removeLast();
+    _notify();
+    return v;
   }
 
   void clear() {
@@ -68,6 +54,7 @@ class RList<V> implements _BaseR {
     _notify();
   }
 
+  /// clear and addAll to List
   void change(List<V> newList) {
     _list.clear();
     _list.addAll(newList);