+++ lib/src/state_with_view_model.dart
@@ -3,14 +3,22 @@ part of view_model_kit;
 /// Use Instead of [State] with [StatefulWidget].
 abstract class StateWithViewModel<T extends StatefulWidget,
     VM extends BaseViewModel> extends State<T> {
-  late final VM viewModel = createViewModel().._setNotify(_notify);
+  /// [viewModel] instance.
+  late final VM viewModel = createViewModel()
+    .._setNotifyAtStatefulWidget(_notifyAtThisWidget);
 
   /// create [viewModel] instance.
   ///
   /// must be override.
   VM createViewModel();
 
-  void _notify() {
+  @override
+  void dispose() {
+    viewModel.dispose();
+    super.dispose();
+  }
+
+  void _notifyAtThisWidget() {
     if (mounted) setState(() {});
   }
 }