+++ lib/view_model_kit.dart
@@ -3,6 +3,15 @@ library view_model_kit;
 import 'package:flutter/widgets.dart';
 
 part 'src/view_model.dart';
+
 part 'src/state_with_view_model.dart';
-part 'src/rx_value.dart';
-part 'src/rx_value_extension.dart';
+
+part 'src/rx/rx_base.dart';
+
+part 'src/rx/rx_value_impl.dart';
+
+part 'src/rx/rx_list_impl.dart';
+
+part 'src/rx/rx_value_extension.dart';
+
+part 'src/select_builder_widget.dart';