+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 0.0.2
+- add R<V>.observe and RList<V>.observe method.
+- add SelectBuilder Widget.
+
 ## 0.0.1
-initial release
+- initial release